Revamped Menus, Levels, AI, and Textures
Changed the glass colour to be clearer.
Added Textures to the keyboard and mouse of the computer.
Added glass to the doors.
Tweaked the size of the Security Guard hitbox.
Added a Demo Level.
Changed the difficulty of all levels to be more gradual.
Changed organisation of menus.
Added Options menu.
Tutorial must now be viewed before playing.
Both the tutorial and options are accessible in the start and pause menus.
The level is now saved.
Progress can now be reset.
Changed the look of the start menu to match the other menus.
Fixed bug where robot sound would still play when hacking or dead.
Fixed bug where the security guard would appear in front of the death screen.
Increased speed and acceleration of the security guard when chasing.